feat: add user authentication and profile management functionality

+++ src/main/java/org/exo/student/controller/StudentController.java
@@ -3,6 +3,7 @@ package org.exo.student.controller;
 import jakarta.validation.Valid;
 import org.exo.student.model.Student;
 import org.exo.student.service.StudentService;
+import org.exo.student.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -12,9 +13,11 @@ import org.springframework.web.bind.annotation.*;
 public class StudentController {
 
     private final StudentService studentService;
+    private final UserService userService;
 
-    public StudentController(StudentService studentService) {
+    public StudentController(StudentService studentService, UserService userService) {
         this.studentService = studentService;
+        this.userService = userService;
     }
 
     @RequestMapping("/")
@@ -25,59 +28,82 @@ public class StudentController {
 
     @GetMapping("/students")
     public String students(@RequestParam(value = "search", required = false) String search, Model model) {
-        if (search != null && !search.isEmpty()) {
-            model.addAttribute("students", studentService.searchStudent(search));
+        if (userService.checkLogin()) {
+            if (search != null && !search.isEmpty()) {
+                model.addAttribute("students", studentService.searchStudent(search));
+            } else {
+                model.addAttribute("students", studentService.getAllStudents());
+            }
+            return "pages/students";
         } else {
-            model.addAttribute("students", studentService.getAllStudents());
+            return "auth/login";
         }
-        return "pages/students";
     }
 
     @GetMapping("/students/register")
     public String register(Model model) {
-        Student student = new Student();
-        model.addAttribute("student", student);
-        return "pages/form";
+        if (userService.checkLogin()) {
+            Student student = new Student();
+            model.addAttribute("student", student);
+            return "pages/form";
+        } else {
+            return "auth/login";
+        }
     }
 
     @PostMapping(value = "/students/register/")
     public String createStudent(@Valid @ModelAttribute("student") Student student, BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            return "pages/form";
-        } else {
-            if (student.getId() != 0) {
-                studentService.updateStudent(student.getId(), student);
+        if (userService.checkLogin()) {
+            if (bindingResult.hasErrors()) {
+                return "pages/form";
             } else {
                 studentService.saveStudent(student);
             }
+            return "redirect:/students";
+        } else {
+            return "auth/login";
         }
-        return "redirect:/students";
     }
 
     @RequestMapping("/students/details/{id}")
     public String details(Model model, @PathVariable("id") int id) {
-        model.addAttribute("student", studentService.getOneStudent(id));
-        return "pages/details";
+        if (userService.checkLogin()) {
+            Student student = studentService.getOneStudent(id);
+            model.addAttribute("student", student);
+            return "pages/details";
+        } else {
+            return "auth/login";
+        }
     }
 
     @PostMapping(value = "/students/delete/{id}")
     public String deleteStudent(@PathVariable("id") int id) {
-        studentService.deleteStudent(id);
-        return "redirect:/students";
+        if (userService.checkLogin()) {
+            studentService.deleteStudent(id);
+            return "redirect:/students";
+        } else {
+            return "auth/login";
+        }
     }
 
     @GetMapping("/students/update/{id}")
     public String showUpdateForm(@PathVariable("id") int id, Model model) {
-        Student student = studentService.getOneStudent(id);
-        model.addAttribute("student", student);
-        return "pages/form";
+        if (userService.checkLogin()) {
+            Student student = studentService.getOneStudent(id);
+            model.addAttribute("student", student);
+            return "pages/update";
+        } else {
+            return "auth/login";
+        }
     }
 
     @PostMapping("/students/update/{id}")
     public String updateStudent(@ModelAttribute Student student) {
-        studentService.saveStudent(student);
-        return "redirect:/students";
+        if (userService.checkLogin()) {
+            studentService.saveStudent(student);
+            return "redirect:/students";
+        } else {
+            return "auth/login";
+        }
     }
-
-
 }
